Removed unused methods and moved misplaced methods
