+++ src/main/java/net/sf/jabref/logic/util/strings/StringUtil.java
@@ -621,4 +621,86 @@ public class StringUtil {
         return result;
     }
 
+    /**
+     * Expand initials, e.g. EH Wissler -> E. H. Wissler or Wissler, EH -> Wissler, E. H.
+     *
+     * @param name
+     * @return The name after expanding initials.
+     */
+    public static String expandAuthorInitials(String name) {
+        String[] authors = name.split(" and ");
+        StringBuilder sb = new StringBuilder();
+        for (int i = 0; i < authors.length; i++) {
+            if (authors[i].contains(", ")) {
+                String[] names = authors[i].split(", ");
+                if (names.length > 0) {
+                    sb.append(names[0]);
+                    if (names.length > 1) {
+                        sb.append(", ");
+                    }
+                }
+                for (int j = 1; j < names.length; j++) {
+                    if (j == 1) {
+                        sb.append(StringUtil.expandAll(names[j]));
+                    } else {
+                        sb.append(names[j]);
+                    }
+                    if (j < (names.length - 1)) {
+                        sb.append(", ");
+                    }
+                }
+    
+            } else {
+                String[] names = authors[i].split(" ");
+                if (names.length > 0) {
+                    sb.append(StringUtil.expandAll(names[0]));
+                }
+                for (int j = 1; j < names.length; j++) {
+                    sb.append(' ');
+                    sb.append(names[j]);
+                }
+            }
+            if (i < (authors.length - 1)) {
+                sb.append(" and ");
+            }
+        }
+    
+        return sb.toString().trim();
+    }
+
+    public static String expandAll(String s) {
+        //System.out.println("'"+s+"'");
+        // Avoid arrayindexoutof.... :
+        if (s.isEmpty()) {
+            return s;
+        }
+        // If only one character (uppercase letter), add a dot and return immediately:
+        if ((s.length() == 1) && Character.isLetter(s.charAt(0)) &&
+                Character.isUpperCase(s.charAt(0))) {
+            return s + ".";
+        }
+        StringBuilder sb = new StringBuilder();
+        char c = s.charAt(0);
+        char d = 0;
+        for (int i = 1; i < s.length(); i++) {
+            d = s.charAt(i);
+            if (Character.isLetter(c) && Character.isUpperCase(c) &&
+                    Character.isLetter(d) && Character.isUpperCase(d)) {
+                sb.append(c);
+                sb.append(". ");
+            } else {
+                sb.append(c);
+            }
+            c = d;
+        }
+        if (Character.isLetter(c) && Character.isUpperCase(c) &&
+                Character.isLetter(d) && Character.isUpperCase(d)) {
+            sb.append(c);
+            sb.append(". ");
+        } else {
+            sb.append(c);
+        }
+        return sb.toString().trim();
+    }
+
 }
