+++ src/test/java/net/sf/jabref/importer/ImportFormatReaderTest.java
@@ -1,0 +1,20 @@
+package net.sf.jabref.importer;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+
+public class ImportFormatReaderTest {
+
+    @Test
+    public void testImportFromFileImportFormatStringOutputPrinter() {
+        fail("Not yet implemented");
+    }
+
+    @Test
+    public void testImportUnknownFormat() {
+        fail("Not yet implemented");
+    }
+
+}
